make network page public, redirect to apps page for dashboard
